+++ ChaosDataService/db_system/DBIndexCursor.h
@@ -38,7 +38,7 @@ namespace chaos {
 				friend class DBDriver;
 			protected:
 				//! the total number of ellement that are found by qury
-				uint32_t element_found;
+				uint64_t element_found;
 				
 				//!
 				uint32_t result_page_dimension;
@@ -59,7 +59,7 @@ namespace chaos {
 				uint32_t getResultPageDimension();
 			
 				//!
-				uint32_t getNumberOfElementFound();
+				uint64_t getNumberOfElementFound();
 				
 				//!
 				void setResultPageDimension(uint32_t);
